fix: variable reuse when values are null/undefined
When calling nested mutations with optional variables, the
`getVarsAndDefMapper_` function in `aliasDocVariables_.ts`
would reuse variables inappropriately.

For example, given a mutation like:

```
mutation updateIssue(
  $id: String!
  $title: String
  $description: String
  $stateId: String
  $priority: Int
  $estimate: Int
  $assigneeId: String
  $projectId: String
  $labelIds: [String!]
) {
  issueUpdate(
    id: $id
    input: {
      title: $title
      description: $description
      stateId: $stateId
      priority: $priority
      estimate: $estimate
      assigneeId: $assigneeId
      projectId: $projectId
      labelIds: $labelIds
    }
  ) {
    success
    issue {
      id
      title
      updatedAt
    }
  }
}
```

When `updateIssue` would be called only with $id and $estimate, the
omitted $title variable would be reused for the value of $stateId,
$priority, $assigneeId, etc. In other words, this would transform to:

```
issueUpdate(
id: $id
input: {
  title: $title
  description: $description
  stateId: $title
  priority: $title
  estimate: $estimate
  assigneeId: $title
  projectId: $title
  labelIds: $title
  }
)
```

This causes an error on the remote service when the input types
don't match. In this case, when $title (a string) is provided to
$priority, the remote GraphQL service gets angry and errors.

The fix is to not re-use variables if their value is `null` or
`undefined`.